make CasmContractClass available in no_std
