+++ crates/cairo-lang-starknet/src/casm_contract_class/from_contract_class.rs
@@ -1,10 +1,6 @@
-#[cfg(test)]
-#[path = "casm_contract_class_test.rs"]
-mod test;
-
 use cairo_felt::Felt252;
 use cairo_lang_casm::assembler::AssembledCairoProgram;
-use cairo_lang_casm::hints::{Hint, PythonicHint};
+use cairo_lang_casm::hints::PythonicHint;
 use cairo_lang_sierra::extensions::array::ArrayType;
 use cairo_lang_sierra::extensions::bitwise::BitwiseType;
 use cairo_lang_sierra::extensions::ec::EcOpType;
@@ -25,7 +21,7 @@ use cairo_lang_sierra_to_casm::compiler::CompilationError;
 use cairo_lang_sierra_to_casm::metadata::{
     calc_metadata, MetadataComputationConfig, MetadataError,
 };
-use cairo_lang_utils::bigint::{deserialize_big_uint, serialize_big_uint, BigUintAsHex};
+use cairo_lang_utils::bigint::BigUintAsHex;
 use cairo_lang_utils::ordered_hash_map::OrderedHashMap;
 use cairo_lang_utils::unordered_hash_set::UnorderedHashSet;
 use convert_case::{Case, Casing};
@@ -34,15 +30,20 @@ use num_bigint::BigUint;
 use num_integer::Integer;
 use num_traits::Signed;
 use once_cell::sync::Lazy;
-use serde::{Deserialize, Serialize};
-use starknet_crypto::{poseidon_hash_many, FieldElement};
 use thiserror::Error;
 
 use crate::allowed_libfuncs::AllowedLibfuncsError;
+use crate::casm_contract_class::{CasmContractEntryPoint, CasmContractEntryPoints};
 use crate::compiler_version::{current_compiler_version_id, current_sierra_version_id, VersionId};
 use crate::contract::starknet_keccak;
 use crate::contract_class::{ContractClass, ContractEntryPoint};
 use crate::felt252_serde::{sierra_from_felt252s, Felt252SerdeError};
+
+use super::CasmContractClass;
+
+#[cfg(test)]
+#[path = "from_contract_class_test.rs"]
+mod test;
 
 /// The expected gas cost of an entrypoint.
 pub const ENTRY_POINT_COST: i32 = 10000;
@@ -81,73 +82,6 @@ pub enum StarknetSierraCompilationError {
          version: {version_of_compiler})"
     )]
     UnsupportedSierraVersion { version_in_contract: VersionId, version_of_compiler: VersionId },
-}
-
-fn skip_if_none<T>(opt_field: &Option<T>) -> bool {
-    opt_field.is_none()
-}
-
-/// Represents a contract in the Starknet network.
-#[derive(Clone, Default, Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct CasmContractClass {
-    #[serde(serialize_with = "serialize_big_uint", deserialize_with = "deserialize_big_uint")]
-    pub prime: BigUint,
-    pub compiler_version: String,
-    pub bytecode: Vec<BigUintAsHex>,
-    pub hints: Vec<(usize, Vec<Hint>)>,
-
-    // Optional pythonic hints in a format that can be executed by the python vm.
-    #[serde(skip_serializing_if = "skip_if_none")]
-    pub pythonic_hints: Option<Vec<(usize, Vec<String>)>>,
-    pub entry_points_by_type: CasmContractEntryPoints,
-}
-impl CasmContractClass {
-    /// Returns the hash value for the compiled contract class.
-    pub fn compiled_class_hash(&self) -> Felt252 {
-        // Compute hashes on each component separately.
-        let external_funcs_hash = self.entry_points_hash(&self.entry_points_by_type.external);
-        let l1_handlers_hash = self.entry_points_hash(&self.entry_points_by_type.l1_handler);
-        let constructors_hash = self.entry_points_hash(&self.entry_points_by_type.constructor);
-        let bytecode_hash = poseidon_hash_many(
-            &self
-                .bytecode
-                .iter()
-                .map(|big_uint| {
-                    FieldElement::from_byte_slice_be(&big_uint.value.to_bytes_be()).unwrap()
-                })
-                .collect_vec(),
-        );
-
-        // Compute total hash by hashing each component on top of the previous one.
-        Felt252::from_bytes_be(
-            &poseidon_hash_many(&[
-                FieldElement::from_byte_slice_be(b"COMPILED_CLASS_V1").unwrap(),
-                external_funcs_hash,
-                l1_handlers_hash,
-                constructors_hash,
-                bytecode_hash,
-            ])
-            .to_bytes_be(),
-        )
-    }
-    /// Returns the hash for a set of entry points.
-    fn entry_points_hash(&self, entry_points: &[CasmContractEntryPoint]) -> FieldElement {
-        let mut entry_point_hash_elements = vec![];
-        for entry_point in entry_points {
-            entry_point_hash_elements.push(
-                FieldElement::from_byte_slice_be(&entry_point.selector.to_bytes_be()).unwrap(),
-            );
-            entry_point_hash_elements.push(FieldElement::from(entry_point.offset));
-            entry_point_hash_elements.push(poseidon_hash_many(
-                &entry_point
-                    .builtins
-                    .iter()
-                    .map(|builtin| FieldElement::from_byte_slice_be(builtin.as_bytes()).unwrap())
-                    .collect_vec(),
-            ));
-        }
-        poseidon_hash_many(&entry_point_hash_elements)
-    }
 }
 
 /// Context for resolving types.
@@ -476,25 +410,4 @@ impl CasmContractClass {
             },
         })
     }
-}
-
-#[derive(Clone, Default, Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct CasmContractEntryPoint {
-    /// A field element that encodes the signature of the called function.
-    #[serde(serialize_with = "serialize_big_uint", deserialize_with = "deserialize_big_uint")]
-    pub selector: BigUint,
-    /// The offset of the instruction that should be called within the contract bytecode.
-    pub offset: usize,
-    // list of builtins.
-    pub builtins: Vec<String>,
-}
-
-#[derive(Clone, Default, Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct CasmContractEntryPoints {
-    #[serde(rename = "EXTERNAL")]
-    pub external: Vec<CasmContractEntryPoint>,
-    #[serde(rename = "L1_HANDLER")]
-    pub l1_handler: Vec<CasmContractEntryPoint>,
-    #[serde(rename = "CONSTRUCTOR")]
-    pub constructor: Vec<CasmContractEntryPoint>,
 }
